Split Conversion into LineResultsGenerator, OutputGenerator
It's weird having ConversionContext and Conversion. This clarifies the intent of each part of the 2-step process:
1. Create temporary line summaries from raw GLS
2. Convert those line summaries into language output
