+++ test/end-to-end/Classes/declarations.js
@@ -4,7 +4,7 @@ class Point {
     public y;
     protected square;
     private name;
-    
+
     constructor(x, y) {
         this.x = x;
         this.y = y;
