+++ test/end-to-end/Classes/declarations.py
@@ -1,6 +1,6 @@
 #
 class Point:
-    
+
     def __init__(self, x, y):
         self.x = x
         self.y = y
