+++ test/end-to-end/Main/main.java
@@ -3,7 +3,7 @@ class Program {
     void sayHello(string name) {
         System.out.println("Hello, " + name + "!");
     }
-    
+
     public static void main(String[] args) {
         sayHello("GLS");
     }
